# frontend/src/openapi/openapi.json
{"openapi":"3.1.0","info":{"title":"DnD helper API","description":"API for generating images for DnD helper app","version":"1.0.0"},"paths":{"/":{"get":{"summary":"Root","operationId":"root__get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}}}},"/image":{"post":{"summary":"Generate Image","operationId":"generate_image_image_post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/ImageRequest"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ImageResponse"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/chat/{thread_id}":{"get":{"summary":"Get Chat","operationId":"get_chat_chat__thread_id__get","parameters":[{"name":"thread_id","in":"path","required":true,"schema":{"type":"string","title":"Thread Id"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ChatResponse"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/chat":{"put":{"summary":"Reply Chat","operationId":"reply_chat_chat_put","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/NewMessageRequest"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ChatResponse"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}},"post":{"summary":"Create Chat","operationId":"create_chat_chat_post","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/NewChatResponse"}}}}}}},"/voice":{"get":{"summary":"Voice","operationId":"voice_voice_get","parameters":[{"name":"text","in":"query","required":true,"schema":{"type":"string","title":"Text"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}}},"components":{"schemas":{"ChatResponse":{"properties":{"messages":{"items":{"$ref":"#/components/schemas/Message"},"type":"array","title":"Messages"}},"type":"object","required":["messages"],"title":"ChatResponse"},"HTTPValidationError":{"properties":{"detail":{"items":{"$ref":"#/components/schemas/ValidationError"},"type":"array","title":"Detail"}},"type":"object","title":"HTTPValidationError"},"ImageRequest":{"properties":{"prompt":{"type":"string","title":"Prompt"}},"type":"object","required":["prompt"],"title":"ImageRequest"},"ImageResponse":{"properties":{"image_url":{"type":"string","title":"Image Url"}},"type":"object","required":["image_url"],"title":"ImageResponse"},"Message":{"properties":{"text":{"type":"string","title":"Text"},"role":{"type":"string","enum":["user","assistant"],"title":"Role"},"type":{"type":"string","enum":["text","image"],"title":"Type"}},"type":"object","required":["text","role","type"],"title":"Message"},"NewChatResponse":{"properties":{"thread_id":{"type":"string","title":"Thread Id"},"messages":{"items":{"$ref":"#/components/schemas/Message"},"type":"array","title":"Messages"}},"type":"object","required":["thread_id","messages"],"title":"NewChatResponse"},"NewMessageRequest":{"properties":{"thread_id":{"type":"string","title":"Thread Id"},"message":{"$ref":"#/components/schemas/Message"}},"type":"object","required":["thread_id","message"],"title":"NewMessageRequest"},"ValidationError":{"properties":{"loc":{"items":{"anyOf":[{"type":"string"},{"type":"integer"}]},"type":"array","title":"Location"},"msg":{"type":"string","title":"Message"},"type":{"type":"string","title":"Error Type"}},"type":"object","required":["loc","msg","type"],"title":"ValidationError"}}}}